refactor(web): move `VariableStoreDictionary` to `variableStore.ts`
